// Web/src/types/define.ts
export enum SidebarType {
    None = 'none',
    About = 'about',
    Blog = 'blog'
}
export enum IconName {
    Error = 'ErrorIcon',
    Home ='HomeIcon',
    About = 'AboutIcon',
    Share = 'ShareIcon',
    Blog = 'BlogIcon',
    Plan = 'PlanIcon',
    Wish = 'WishIcon',
    Random = 'RandomIcon',
    Skill = 'SkillIcon'
}